requirements.txt
absl-py==2.0.0
actionlib==1.14.0
ale-py==0.8.1
angles==1.9.13
asttokens==2.4.1
attrs==24.2.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
base_local_planner==1.17.3
bleach==6.1.0
blessed==1.20.0
bondpy==1.8.6
box2d-py==2.3.5
cachetools==5.4.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.7.4
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
cloudpickle==3.0.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cvxpy==1.5.2
cycler==0.12.1
Cython==3.0.10
debugpy @ file:///croot/debugpy_1690905042057/work
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
dynamic-reconfigure==1.7.3
ecos==2.0.14
empy==3.3.4
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
envpool==0.8.4
executing==2.0.1
faiss-gpu==1.7.2
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filterpy==1.4.5
fonttools==4.51.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.65.4
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.11.0
idna==3.7
image-geometry==1.16.2
imageio==2.35.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
interactive-markers==1.12.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lap==0.4.0
laser_geometry==1.6.7
loguru==0.7.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
message-filters==1.17.0
mistune==3.0.2
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
ninja==1.11.1.1
numpy==1.22.0
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.560.30
oauthlib==3.2.2
onnx==1.13.0
onnx-simplifier==0.4.10
opencv-python==4.8.1.78
optree==0.13.1
osqp==0.6.7.post1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pygame==2.6.1
pyglet==2.0.18
Pygments==2.17.2
pyparsing==3.1.2
pyrealsense2==2.54.2.5684
pyrvo2==0.0.0
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
qdldl==0.1.7.post4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rich==13.7.0
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rpds-py==0.20.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.25
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.6
sensor-msgs==1.13.1
sentry-sdk==2.18.0
setproctitle==1.3.4
Shimmy==1.3.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
stable_baselines3==2.3.2
stack-data==0.6.3
swig==4.2.1.post0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tinycss2==1.3.0
topic-tools==1.17.0
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transforms3d==0.4.2
types-protobuf==5.28.3.20241030
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
xacro==1.14.18
yacs==0.1.8
yarg==0.1.9
-e git+https://github.com/Megvii-BaseDetection/YOLOX.git@ac58e0a5e68e57454b7b9ac822aced493b553c53#egg=yolox
zipp==3.18.1
